+++ src/main/java/net/spell_engine/api/item/trinket/SpellBookItem.java
@@ -1,14 +1,27 @@
 package net.spell_engine.api.item.trinket;
 
 import net.minecraft.item.Item;
-import net.minecraft.item.ItemConvertible;
+import net.minecraft.item.ItemStack;
+import net.minecraft.sound.SoundEvent;
 import net.minecraft.util.Identifier;
-import net.spell_engine.api.item.SpellEngineItemTags;
 
-public interface SpellBookItem extends ItemConvertible {
-    Identifier getPoolId();
+public class SpellBookItem extends Item implements ISpellBookItem {
+    public static final Identifier EQUIP_SOUND_ID = Identifier.of("spell_engine", "spellbook_equip");
+    public static final SoundEvent EQUIP_SOUND = SoundEvent.of(EQUIP_SOUND_ID);
 
-    static boolean isSpellBook(Item item) {
-        return item instanceof SpellBookItem || item.getRegistryEntry().isIn(SpellEngineItemTags.SPELL_BOOK);
+    private final Identifier poolId;
+    public SpellBookItem(Identifier poolId, Settings settings) {
+        super(settings);
+        this.poolId = poolId;
+    }
+
+    @Override
+    public boolean isEnchantable(ItemStack stack) {
+        return false;
+    }
+
+    @Override
+    public Identifier getPoolId() {
+        return poolId;
     }
 }