+++ src/main/java/net/spell_engine/fabric/FabricMod.java
@@ -10,7 +10,7 @@ import net.minecraft.registry.Registry;
 import net.minecraft.util.Identifier;
 import net.spell_engine.SpellEngineMod;
 import net.spell_engine.api.effect.RemoveOnHit;
-import net.spell_engine.compat.TrinketsCompat;
+import net.spell_engine.compat.trinkets.TrinketsCompat;
 import net.spell_engine.entity.SpellCloud;
 import net.spell_engine.entity.SpellProjectile;
 import net.spell_engine.item.SpellEngineItems;
@@ -44,10 +44,9 @@ public class FabricMod implements ModInitializer {
     public void onInitialize() {
         SpellEngineMod.init();
         SpellEngineMod.registerSpellBinding();
-        SpellEngineItems.register();
         SoundHelper.registerSounds();
-
         TrinketsCompat.init();
+        SpellEngineItems.register();
 
         ServerLivingEntityEvents.ALLOW_DAMAGE.register((entity, source, amount) -> {
             var attacker = source.getAttacker();